Increase script flexibility by adding a configurable duration between checks.
The commit introduces a new configuration variable `DURATION_BETWEEN_CHECKS` in the file `config.conf`. This variable allows the user to set the duration in seconds between each check in the manager script (`POCIDFBIManager.sh`).

The default value for `DURATION_BETWEEN_CHECKS` is 10 seconds. However, the user can modify this value in the `config.conf` file by setting the desired duration.