feat: Add option to disable cron setup
This adds the functionality to disable cron setup when running the install script. The user can now pass the "-n" option to prevent the cron task from being created. If the cron task does not exist and the user does not want to skip it, a message will be displayed with the command to manually add the cron task to the crontab.